@@ -30,13 +30,13 @@
 </style>
 
 <script lang="ts">
-import Vue from "vue";
+import { defineComponent } from "vue";
 import Store from "../store";
 
-export default Vue.extend({
+export default defineComponent({
   name: "Home",
   data: () => ({
-    show: true
+    show: true,
   }),
   async mounted() {
     document.title = "YaMusic";
@@ -54,7 +54,8 @@ export default Vue.extend({
     }
 
     webview.addEventListener("dom-ready", async () => {
-      await webview.insertCSS(`              
+      try {
+        await webview.insertCSS(`              
         /* Overhead */
         .d-overhead {
           display: none !important;
@@ -91,59 +92,66 @@ export default Vue.extend({
         }
         
         /* Custom */
-        # CSS class dor div.centerblock-wrapper
         .centerblock-wrapper__no_margin {
           margin-right: 0px !important;
         }
       `);
 
-      // Hide sidebar
-      await webview.executeJavaScript(`
-      function hideSidebar(){
-        const centralBlockDOM = document.querySelector('div.centerblock-wrapper');
-        const sidebarDOM = document.querySelector('.sidebar');
-        
-        centralBlockDOM.style.marginRight = '0px';
-
-        const observer = new MutationObserver(function(mutations) {    
-          const trackSidebarDOM = document.querySelector('.sidebar-cont.sidebar-cont_shown.deco-pane'); 
-            
-          if (trackSidebarDOM) {
-              centralBlockDOM.style.marginRight = '';
-              observer.disconnect();
-          } else {
-              centralBlockDOM.style.marginRight = '0px';
+        // Hide sidebar
+        await webview.executeJavaScript(`
+        function hideSidebar(){
+          const centralBlockDOM = document.querySelector('div.centerblock-wrapper');
+          const sidebarDOM = document.querySelector('.sidebar');
+          
+          if (!centralBlockDOM || !sidebarDOM) {
+            setTimeout(hideSidebar, 500);
+            return;
           }
 
-          setTimeout(startObserving,1000);
-        });
+          centralBlockDOM.style.marginRight = '0px';
 
-        const startObserving = function() {
-            observer.observe(sidebarDOM, { 
-                childList: true,
-                subtree: true
-            });
+          const observer = new MutationObserver(function(mutations) {    
+            const trackSidebarDOM = document.querySelector('.sidebar-cont.sidebar-cont_shown.deco-pane'); 
+              
+            if (trackSidebarDOM) {
+                centralBlockDOM.style.marginRight = '';
+                observer.disconnect();
+            } else {
+                centralBlockDOM.style.marginRight = '0px';
+            }
+
+            setTimeout(startObserving,1000);
+          });
+
+          const startObserving = function() {
+              observer.observe(sidebarDOM, { 
+                  childList: true,
+                  subtree: true
+              });
+          }
+
+          startObserving();
+        }
+        
+        function innerRoute(location){
+          let tmpDom = document.createElement('a');
+          tmpDom.id = 'innerRoute'
+          tmpDom.setAttribute('href', location);
+          
+          document.body.appendChild(tmpDom);       
+          document.querySelector('#innerRoute').click();
         }
 
-        startObserving();
+        // Init
+        hideSidebar();
+        `);
+      } catch (err) {
+        console.error("webview dom-ready handler failed", err);
       }
-      
-      function innerRoute(location){
-        let tmpDom = document.createElement('a');
-        tmpDom.id = 'innerRoute'
-        tmpDom.setAttribute('href', location);
-        
-        document.body.appendChild(tmpDom);       
-        document.querySelector('#innerRoute').click();
-      }
-
-      // Init
-      hideSidebar();
-      `);
 
       // did-stop-loading may not work
       loadstop();
     });
-  }
+  },
 });
 </script>
