@@ -1,23 +1,23 @@
 "use strict";
 
-import * as path from 'path';
+import * as path from "path";
 import { app, protocol, session, BrowserWindow } from "electron";
 import { createProtocol } from "vue-cli-plugin-electron-builder/lib";
 import installExtension, { VUEJS_DEVTOOLS } from "electron-devtools-installer";
+import { autoUpdater } from "electron-updater";
 
-import store from './electron/store';
-import ctxMenu from './electron/menu';
-import globalShortcut from './electron/globalShortcut';
-import notification from './electron/notification';
-import proxy from './electron/proxy';
+import store from "./electron/store";
+import ctxMenu from "./electron/menu";
+import globalShortcut from "./electron/globalShortcut";
+import notification from "./electron/notification";
+import proxy from "./electron/proxy";
 
 const isDevelopment = process.env.NODE_ENV !== "production";
 const _defaultHeight = 620;
 const _defaultWidth = 820;
 
-// Scheme must be registered before the app is ready
 protocol.registerSchemesAsPrivileged([
-  { scheme: "app", privileges: { secure: true, standard: true } }
+  { scheme: "app", privileges: { secure: true, standard: true } },
 ]);
 
 let win: BrowserWindow | null = null;
@@ -27,68 +27,59 @@ if (!appRunning) {
   app.quit();
 }
 
-app.on('second-instance', () => {
+app.on("second-instance", () => {
   if (win) {
     if (win.isMinimized() || !win.isVisible()) {
       win.restore();
     }
     win.focus();
   }
-})
+});
 
 async function createWindow() {
-  const lastWindowState = store.get('lastWindowState');
+  const lastWindowState = store.get("lastWindowState");
 
-  // Create the browser window.
   win = new BrowserWindow({
-    title: 'YaMusic',
+    title: "YaMusic",
     x: lastWindowState ? lastWindowState.x : 100,
     y: lastWindowState ? lastWindowState.y : 100,
     height: lastWindowState ? lastWindowState.height : _defaultHeight,
     width: lastWindowState ? lastWindowState.width : _defaultWidth,
-    icon: path.join(__static, 'icon.png'),
-    titleBarStyle: 'hiddenInset',
+    icon: path.join(__static, "icon.png"),
     minHeight: _defaultHeight,
     minWidth: _defaultWidth,
     autoHideMenuBar: true,
-    backgroundColor: '#ECEFF1', //TODO this
+    backgroundColor: "#ECEFF1",
     webPreferences: {
-      webviewTag: true, // Security warning since Electron 10
+      webviewTag: true,
       zoomFactor: 1.0,
       enableRemoteModule: true,
-      // Use pluginOptions.nodeIntegration, leave this alone
-      // See nklayman.github.io/vue-cli-plugin-electron-builder/guide/security.html#node-integration for more info
-      // (process.env.ELECTRON_NODE_INTEGRATION as unknown) as boolean
       nodeIntegration: true,
-    }
+      contextIsolation: false,
+    },
   });
 
   if (process.env.WEBPACK_DEV_SERVER_URL) {
-    // Load the url of the dev server if in development mode
     await win.loadURL(process.env.WEBPACK_DEV_SERVER_URL as string);
     if (!process.env.IS_TEST) win.webContents.openDevTools();
   } else {
-    const { autoUpdater } = require("electron-updater");
     createProtocol("app");
-    // Load the index.html when not in development
     win.loadURL("app://./index.html");
     autoUpdater.checkForUpdatesAndNotify();
   }
 
-  win.on('close', e => {
-    if (!store.get('quit?')) {
+  win.on("close", (e) => {
+    if (!store.get("quit?")) {
       e.preventDefault();
     }
 
     if (win) {
       switch (process.platform) {
-        case 'win32':
+        case "win32":
+        case "linux":
           win.hide();
           break;
-        case 'linux':
-          win.hide();
-          break;
-        case 'darwin':
+        case "darwin":
           app.hide();
           break;
         default:
@@ -96,31 +87,25 @@ async function createWindow() {
     }
   });
 
-  return win
+  return win;
 }
 
-// Quit when all windows are closed.
 app.on("window-all-closed", () => {
-  // On macOS it is common for applications and their menu bar
-  // to stay active until the user quits explicitly with Cmd + Q
-
   if (process.platform !== "darwin") {
     app.quit();
   }
 });
 
 app.on("activate", () => {
-  // On macOS it's common to re-create a window in the app when the
-  // dock icon is clicked and there are no other windows open.
   if (BrowserWindow.getAllWindows().length === 0) createWindow();
 });
 
-// This method will be called when Electron has finished
-// initialization and is ready to create browser windows.
-// Some APIs can only be used after this event occurs.
 app.on("ready", async () => {
-  // Skip auto-installing Vue Devtools to avoid extension manifest warnings on old Electron/Chrome
-  if (isDevelopment && !process.env.IS_TEST && process.env.VUE_DEVTOOLS === "1") {
+  if (
+    isDevelopment &&
+    !process.env.IS_TEST &&
+    process.env.VUE_DEVTOOLS === "1"
+  ) {
     try {
       await installExtension(VUEJS_DEVTOOLS);
     } catch (e) {
@@ -134,55 +119,47 @@ app.on("ready", async () => {
   globalShortcut(win, app);
   await proxy();
 
-  session.fromPartition('persist:webviewsession').webRequest.onBeforeRequest({ urls: ['*://*/*'] }, (details: any, callback) => {    
-    if (details.url.includes("awaps.yandex.net") ||
-      details.url.includes("vh-bsvideo-converted") ||
-      details.url.includes("get-video-an") ||
-      details.url.includes("an.yandex.ru\/vmap") ||
-      details.url.includes("yandex.ru\/an") ||
-      details.url.includes("mc.yandex.ru\/watch") ||
-      details.url.includes("strm.yandex.ru") ||
-      details.url.includes("mc.yandex.ru\/clmap") ||
-      details.url.includes("yandex.ru\/ads")){
+  session
+    .fromPartition("persist:webviewsession")
+    .webRequest.onBeforeRequest(
+      { urls: ["*://*/*"] },
+      (details: any, callback) => {
+        if (
+          details.url.includes("awaps.yandex.net") ||
+          details.url.includes("vh-bsvideo-converted") ||
+          details.url.includes("get-video-an") ||
+          details.url.includes("an.yandex.ru/vmap") ||
+          details.url.includes("yandex.ru/an") ||
+          details.url.includes("mc.yandex.ru/watch") ||
+          details.url.includes("strm.yandex.ru") ||
+          details.url.includes("mc.yandex.ru/clmap") ||
+          details.url.includes("yandex.ru/ads")
+        ) {
+          return { cancel: true };
+        }
+        if (details.url.includes("start?__t")) {
+          notification(win);
+        }
 
-      // Skip advertising
-      return {
-        cancel: true
+        callback({});
       }
-    }
-    if (details.url.includes("start?__t")) {      
-      // Notification for next sing
-      notification(win);
-    }
-
-    callback({});
-  })
+    );
 });
 
-app.on('before-quit', () => {
-  store.set('quit?', true);
+app.on("before-quit", () => {
+  store.set("quit?", true);
   if (win) {
     if (!win.isFullScreen()) {
-      store.set('lastWindowState', win.getBounds());
+      store.set("lastWindowState", win.getBounds());
     }
 
-    store.set('lastApp', win.getTitle());
+    store.set("lastApp", win.getTitle());
   }
 });
 
-// app.on('web-contents-created', function (webContentsCreatedEvent, contents) {
-//   if (contents.getType() === 'webview') {
-//     contents.on('new-window', function (newWindowEvent, url) {      
-//       win?.loadURL(url)
-//       newWindowEvent.preventDefault();
-//     });
-//   }
-// });
-
-// Exit cleanly on request from parent process in development mode.
 if (isDevelopment) {
   if (process.platform === "win32") {
-    process.on("message", data => {
+    process.on("message", (data) => {
       if (data === "graceful-exit") {
         app.quit();
       }
