@@ -5,12 +5,12 @@ Vue.use(Vuex);
 
 export default new Vuex.Store({
   state: {
-    loading: false
+    loading: false,
   },
   mutations: {
-    loadingSetFalse: state => (state.loading = false),
-    loadingSetTrue: state => (state.loading = true)
+    loadingSetFalse: (state) => (state.loading = false),
+    loadingSetTrue: (state) => (state.loading = true),
   },
   actions: {},
-  modules: {}
+  modules: {},
 });
